Adding option to set max players in example

+++ lib/LinkWireless.h
@@ -921,7 +921,7 @@ class LinkWireless {
   void addConfirmations() {  // (irq only)
     if (state == SERVING) {
       addData(buildConfirmationHeader(0));
-      for (u32 i = 0; i < config.maxPlayers - 1; i++)
+      for (u32 i = 0; i < (u32)(config.maxPlayers - 1); i++)
         addData(sessionState.lastPacketIdFromClients[1 + i]);
     } else {
       addData(buildConfirmationHeader(sessionState.currentPlayerId));
@@ -936,7 +936,8 @@ class LinkWireless {
     bool isServerConfirmation = confirmation.playerId == 0;
 
     if (isServerConfirmation) {
-      if (state != CONNECTED || confirmation.dataSize != config.maxPlayers - 1)
+      if (state != CONNECTED ||
+          confirmation.dataSize != (u32)(config.maxPlayers - 1))
         return false;
 
       sessionState.lastConfirmationFromServer =
@@ -951,7 +952,7 @@ class LinkWireless {
           confirmationData;
 
       u32 min = 0xffffffff;
-      for (u32 i = 0; i < config.maxPlayers - 1; i++) {
+      for (u32 i = 0; i < (u32)(config.maxPlayers - 1); i++) {
         u32 confirmationData = sessionState.lastConfirmationFromClients[1 + i];
         if (confirmationData > 0 && confirmationData < min)
           min = confirmationData;
@@ -1347,7 +1348,7 @@ class LinkWireless {
 
   bool timeout(u32 limit, u32& lines, u32& vCount) {
     if (REG_VCOUNT != vCount) {
-      lines += std::max((s32)REG_VCOUNT - (s32)vCount, 0);
+      lines += std::max((int)REG_VCOUNT - (int)vCount, 0);
       vCount = REG_VCOUNT;
     }
 
@@ -1360,7 +1361,7 @@ class LinkWireless {
 
     while (count < verticalLines) {
       if (REG_VCOUNT != vCount) {
-        count += std::max((s32)REG_VCOUNT - (s32)vCount, 0);
+        count += std::max((int)REG_VCOUNT - (int)vCount, 0);
         vCount = REG_VCOUNT;
       }
     };
